refactor: reorganize code into modules
