+++ TODO.md
@@ -1,0 +1,214 @@
+# Eliza Framework Improvement Plan
+
+## High Priority
+
+### 1. Code Organization & Architecture
+
+- [x] Refactor index.ts into modular components:
+  - [x] Create `src/config` for configuration management
+  - [x] Create `src/clients` for client interface management
+  - [x] Create `src/database` for database initialization and management
+  - [x] Create `src/cache` for cache system management
+  - [x] Create `src/chat` for chat interface management
+- [ ] Implement dependency injection system
+- [ ] Create service layer abstractions:
+  - [ ] Extract runtime initialization into a separate service
+  - [ ] Create interfaces for all services
+  - [ ] Implement service factory pattern
+- [ ] Standardize error handling across modules
+- [ ] Improve module organization:
+  - [ ] Add barrel exports (index.ts) for each module
+  - [ ] Create shared types module
+  - [ ] Add module documentation
+  - [ ] Implement module-level error handling
+
+### 2. Error Handling & Logging
+
+- [ ] Implement structured logging system:
+  - [ ] Add log levels (DEBUG, INFO, WARN, ERROR)
+  - [ ] Add request/response logging
+  - [ ] Add performance logging
+- [ ] Create custom error types:
+  - [ ] ClientError for client-related issues
+  - [ ] DatabaseError for database operations
+  - [ ] ConfigurationError for config issues
+  - [ ] ValidationError for input validation
+- [ ] Add error recovery mechanisms:
+  - [ ] Automatic reconnection for clients
+  - [ ] Fallback strategies for failed operations
+  - [ ] Circuit breakers for external services
+
+### 3. Type Safety
+
+- [ ] Enable strict TypeScript configuration:
+  - [ ] Set `strict: true` in tsconfig.json
+  - [ ] Set `noImplicitAny: true`
+  - [ ] Enable strict null checks
+- [ ] Add comprehensive type definitions:
+  - [ ] Create types for all configuration options
+  - [ ] Add interfaces for client responses
+  - [ ] Define types for plugin system
+- [ ] Add type guards and assertions
+- [ ] Document type usage patterns
+
+## Medium Priority
+
+### 4. Testing Infrastructure
+
+- [ ] Set up testing framework:
+  - [ ] Add Jest/Mocha for unit testing
+  - [ ] Configure test runners and reporters
+- [ ] Create test suites:
+  - [ ] Unit tests for core functionality
+  - [ ] Integration tests for clients
+  - [ ] E2E tests for critical paths
+- [ ] Add test utilities:
+  - [ ] Mock implementations
+  - [ ] Test data generators
+  - [ ] Test helpers
+- [ ] Set up CI/CD pipeline for tests
+
+### 5. Documentation
+
+- [ ] Add comprehensive JSDoc comments:
+  - [ ] Document all public APIs
+  - [ ] Add examples in comments
+  - [ ] Document configuration options
+- [ ] Create API documentation:
+  - [ ] Generate API docs from JSDoc
+  - [ ] Add usage examples
+  - [ ] Document error handling
+- [ ] Add architecture documentation:
+  - [ ] System overview
+  - [ ] Component interactions
+  - [ ] Data flow diagrams
+- [ ] Create plugin development guide
+
+### 6. Configuration Management
+
+- [x] Create configuration manager:
+  - [x] Centralize config loading
+  - [x] Add environment-specific configs
+  - [x] Add config validation
+- [x] Add configuration schema:
+  - [x] Document all options
+  - [x] Add type definitions
+  - [x] Add validation rules
+- [ ] Implement secure secrets management:
+  - [ ] Add encryption for sensitive data
+  - [ ] Implement key rotation
+  - [ ] Add secrets validation
+- [ ] Add configuration hot reloading
+
+## Lower Priority
+
+### 7. Security
+
+- [ ] Implement input validation:
+  - [ ] Add request validation
+  - [ ] Sanitize user inputs
+  - [ ] Validate configuration values
+- [ ] Add rate limiting:
+  - [ ] Per-client rate limits
+  - [ ] API endpoint rate limits
+  - [ ] Error handling for rate limits
+- [ ] Implement security headers
+- [ ] Add authentication mechanisms:
+  - [ ] API authentication
+  - [ ] Client authentication
+  - [ ] Plugin authentication
+
+### 8. Performance
+
+- [ ] Implement caching strategies:
+  - [ ] Response caching
+  - [ ] Configuration caching
+  - [ ] Resource caching
+- [ ] Add connection pooling:
+  - [ ] Database connection pools
+  - [ ] Client connection management
+- [ ] Optimize database operations:
+  - [ ] Add indexes
+  - [ ] Optimize queries
+  - [ ] Add query caching
+- [ ] Implement request batching
+
+### 9. Development Experience
+
+- [ ] Add development tooling:
+  - [ ] ESLint configuration
+  - [ ] Prettier setup
+  - [ ] Git hooks
+- [ ] Improve build process:
+  - [ ] Add build optimizations
+  - [ ] Reduce build time
+  - [ ] Add build caching
+- [ ] Add debugging capabilities:
+  - [ ] Source maps
+  - [ ] Debug configurations
+  - [ ] Logging utilities
+- [ ] Create development documentation
+
+### 10. Monitoring & Observability
+
+- [ ] Add metrics collection:
+  - [ ] Performance metrics
+  - [ ] Error rates
+  - [ ] Usage statistics
+- [ ] Implement logging aggregation:
+  - [ ] Centralized logging
+  - [ ] Log analysis tools
+  - [ ] Log retention policies
+- [ ] Create monitoring dashboards:
+  - [ ] System health
+  - [ ] Performance metrics
+  - [ ] Error tracking
+- [ ] Add alerting system:
+  - [ ] Error alerts
+  - [ ] Performance alerts
+  - [ ] Usage alerts
+
+## Module-Specific Improvements
+
+### Config Module
+
+- [ ] Add support for multiple character configurations
+- [ ] Implement configuration inheritance
+- [ ] Add configuration validation error messages
+- [ ] Add configuration schema versioning
+
+### Database Module
+
+- [ ] Add migration system
+- [ ] Implement query builder
+- [ ] Add database connection pooling
+- [ ] Add database transaction support
+
+### Cache Module
+
+- [ ] Add cache invalidation strategies
+- [ ] Implement cache size limits
+- [ ] Add cache statistics
+- [ ] Support distributed caching
+
+### Clients Module
+
+- [ ] Add client health checks
+- [ ] Implement client reconnection strategies
+- [ ] Add client event system
+- [ ] Support custom client implementations
+
+### Chat Module
+
+- [ ] Add support for multiple chat interfaces
+- [ ] Implement chat history
+- [ ] Add chat command system
+- [ ] Support rich media messages
+
+## Notes
+
+- Each task should be implemented incrementally
+- Maintain backward compatibility
+- Add tests for new features
+- Update documentation as changes are made
+- Consider performance implications
